Completed Draw FaceMethod
draws faces to the cube and each face is a customizable char. Only bug is that faces are drawn in same order every time and when the front face is behind the back face it will still draw to the front, I will add normals and sorting them in the future.